+++ android/Bottari/presentation/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:windowSoftInputMode="adjustNothing"
             android:launchMode="singleTask" />
         <activity
-            android:name=".view.checklist.ChecklistActivity"
+            android:name=".view.checklist.personal.ChecklistActivity"
             android:exported="false" />
         <activity
             android:name=".view.edit.personal.PersonalBottariEditActivity"
@@ -37,7 +37,7 @@
             android:name=".view.template.TemplateActivity"
             android:exported="false" />
         <activity
-            android:name=".view.team.checklist.TeamChecklistActivity"
+            android:name=".view.checklist.team.TeamChecklistActivity"
             android:exported="false" />
 
         <receiver android:name=".receiver.AlarmReceiver" />
