+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/model/TeamChecklistCategoryUiModel.kt
@@ -1,6 +1,6 @@
 package com.bottari.presentation.model
 
-import com.bottari.presentation.view.team.checklist.checklist.ChecklistCategory
+import com.bottari.presentation.view.checklist.team.checklist.ChecklistCategory
 
 data class TeamChecklistCategoryUiModel(
     val category: ChecklistCategory,
