--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/ChecklistUiEvent.kt
@@ -1,7 +1,0 @@
-package com.bottari.presentation.view.checklist
-
-sealed interface ChecklistUiEvent {
-    data object FetchChecklistFailure : ChecklistUiEvent
-
-    data object CheckItemFailure : ChecklistUiEvent
-}
