--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/main/listener/OnChecklistItemClickListener.kt
@@ -1,5 +1,0 @@
-package com.bottari.presentation.view.checklist.main.listener
-
-fun interface OnChecklistItemClickListener {
-    fun onClick(itemId: Long)
-}
