+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/personal/ChecklistUiEvent.kt
@@ -1,0 +1,7 @@
+package com.bottari.presentation.view.checklist.personal
+
+sealed interface ChecklistUiEvent {
+    data object FetchChecklistFailure : ChecklistUiEvent
+
+    data object CheckItemFailure : ChecklistUiEvent
+}
