+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/personal/main/listener/OnChecklistItemClickListener.kt
@@ -1,0 +1,5 @@
+package com.bottari.presentation.view.checklist.personal.main.listener
+
+fun interface OnChecklistItemClickListener {
+    fun onClick(itemId: Long)
+}
