+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/team/checklist/ChecklistCategory.kt
@@ -1,0 +1,9 @@
+package com.bottari.presentation.view.checklist.team.checklist
+
+enum class ChecklistCategory(
+    val title: String,
+) {
+    SHARED("공통"),
+    ASSIGNED("담당"),
+    PERSONAL("개인"),
+}
