+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/team/checklist/TeamChecklistUiEvent.kt
@@ -1,0 +1,7 @@
+package com.bottari.presentation.view.checklist.team.checklist
+
+sealed interface TeamChecklistUiEvent {
+    data object FetchChecklistFailure : TeamChecklistUiEvent
+
+    data object CheckItemFailure : TeamChecklistUiEvent
+}
