+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/checklist/team/checklist/adapter/OnTeamChecklistItemClickListener.kt
@@ -1,0 +1,5 @@
+package com.bottari.presentation.view.checklist.team.checklist.adapter
+
+fun interface OnTeamChecklistItemClickListener {
+    fun onClick(itemId: Int)
+}
