--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/team/checklist/TeamStatusFragment.kt
@@ -1,20 +1,0 @@
-package com.bottari.presentation.view.team.checklist
-
-import android.os.Bundle
-import android.view.View
-import com.bottari.presentation.common.base.BaseFragment
-import com.bottari.presentation.databinding.FragmentTeamStatusBinding
-
-class TeamStatusFragment : BaseFragment<FragmentTeamStatusBinding>(FragmentTeamStatusBinding::inflate) {
-    override fun onViewCreated(
-        view: View,
-        savedInstanceState: Bundle?,
-    ) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance(): TeamStatusFragment = TeamStatusFragment()
-    }
-}
