--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/team/checklist/checklist/ChecklistCategory.kt
@@ -1,9 +1,0 @@
-package com.bottari.presentation.view.team.checklist.checklist
-
-enum class ChecklistCategory(
-    val title: String,
-) {
-    SHARED("공통"),
-    ASSIGNED("담당"),
-    PERSONAL("개인"),
-}
