--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/team/checklist/checklist/TeamChecklistUiEvent.kt
@@ -1,7 +1,0 @@
-package com.bottari.presentation.view.team.checklist.checklist
-
-sealed interface TeamChecklistUiEvent {
-    data object FetchChecklistFailure : TeamChecklistUiEvent
-
-    data object CheckItemFailure : TeamChecklistUiEvent
-}
