--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/team/checklist/checklist/TeamChecklistUiState.kt
@@ -1,10 +1,0 @@
-package com.bottari.presentation.view.team.checklist.checklist
-
-import com.bottari.presentation.model.TeamChecklistItemUiModel
-import com.bottari.presentation.model.TeamChecklistRowUiModel
-
-data class TeamChecklistUiState(
-    val isLoading: Boolean = false,
-    val items: List<TeamChecklistItemUiModel> = emptyList(),
-    val expandableItems: List<TeamChecklistRowUiModel> = emptyList(),
-)
