--- android/Bottari/presentation/src/main/java/com/bottari/presentation/view/team/checklist/checklist/adapter/OnTeamChecklistItemClickListener.kt
@@ -1,5 +1,0 @@
-package com.bottari.presentation.view.team.checklist.checklist.adapter
-
-fun interface OnTeamChecklistItemClickListener {
-    fun onClick(itemId: Int)
-}
