+++ android/Bottari/presentation/src/main/res/layout/activity_checklist.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.checklist.ChecklistActivity">
+    tools:context=".view.checklist.personal.ChecklistActivity">
 
     <Toolbar
         android:id="@+id/toolbar_bottari"
