+++ android/Bottari/presentation/src/main/res/layout/fragment_member_status.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.team.checklist.MemberStatusFragment">
+    tools:context=".view.checklist.team.MemberStatusFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
