+++ android/Bottari/presentation/src/main/res/layout/fragment_team_checklist.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.team.checklist.checklist.TeamChecklistFragment">
+    tools:context=".view.checklist.team.checklist.TeamChecklistFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_checklist"
