+++ android/Bottari/presentation/src/main/res/layout/fragment_team_status.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.team.checklist.TeamStatusFragment">
+    tools:context=".view.checklist.team.TeamStatusFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
